Restructuring code mostly, explained in description
Removed a few features
-Level start / end
-working timer
As I am reworking the Level data and have begun working on loading it from a file.

Also tuned A lot of values regarding gameplay.

NOTE: Linux driver, gamepad inputs are different depending on if the controller is connected via USB vs BLUETOOTH 
thread for bug: 
https://github.com/paroj/xpad/issues/53